rename(admin): 잘못된 폴더 이름 변경
